Copy from-release pxeboot-update script to deployed /etc
During the major release deployment for multinode systems
there will be a timeframe where hosts will be partially
upgraded to to-release and hosts still not upgraded, running
from-release, and if the need comes to lock/unlock these
hosts in this period, the controllers must have the
pxeboot-update-<release>.sh script for both release, otherwise
the unlock procedure can fail due to these missing scripts.

This commit ensures that during the controllers deployment,
which are the first hosts to be upgraded, that both scripts
for from/to-release are put in /etc so that the failure
mentioned previously is avoided.

This commit also improves the logging around the agent hook
that manages the pxe files, to log the error messages accordingly
when exceptions raise.

Test Plan
PASS: Storage - in major release deploy, in a scenario where the
      system is partially in from and partially in to-release,
      lock/unlock storage and compute hosts and verify no failures
      (given the lock/unlock respect other existing rules for USM)

Story: 2010676
Task: 51241

Change-Id: I01b088904177b784ba2153a21c0c581f0ae111d1
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>