Reload release data during major deployment
It was observed that after applying RR patches to a 24.09 system and
upgrading it to 25.03, the local storage value for 'release_data' is
outdated, containing the old releases. This causes the
is_major_release_deployment() method to incorrectly returns False,
preventing the creation of a new UUID configuration. Instead, the
default 'install' configuration is applied during the unlock, raising
a 250.001 alarm.

This commit adds a reload_release_data() call in the __init__ method
of the PatchMessageAgentInstallResp class, to clean up the old release
data in the local storage.

Test Plan:
PASS: AIO-DX and STD deploy complete.
PASS: Successfully applied RR patch-1 and patch-2 via USM.
PASS: Software upload from 24.09 to 25.03. Verify that a new uuid
configuration is created and applied during the unlock of
controller-1, and that no 250.001 alarm is raised once it is back
online.

Closes-bug: 2086134

Change-Id: Ifc73cafd0e8208e42806ebceb4f35068f128a19f
Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>