Fix failed state during abort
During 'software deploy abort', the software loop through all hosts
and set it to 'rollback-deployed' or rollback-pending' state.
Every time the host state is changed, a listener function updates the
deployment state accordingly based on the current state of all hosts.
If at least one host is in a failed state, the deployment is also set
to 'failed'.

Since this listener is called on every host state change and the
failed host can be the last in the abort list to be updated to
'rollback-pending', the deployment can be set to 'failed' multiple
times.

This commit fixes it by checking if the deployment is already in a
failed state, if yes, just log a message, if not update the state.

Test-Plan-DX:
PASS: Fail a patch during deploy host, abort it, and
      check if all hosts are set to the right state.

Story: 2010676
Task: 51252

Change-Id: I06ff98efe43cde80b37bb1200a355d97ccd64b64
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>