Sync /etc/update.d between controllers
This folder stores all patch activation scripts and it is created by
'deploy start' in the current active controller. 'deploy activate'
runs all scripts inside this folder. But if this second command is
executed from a different controller, the folder with all scripts
will be not present there.

This commit adds a code to sync the /etc/update.d between the
controllers.

Test-plan:
PASS: Apply a patch, running 'deploy start' and
      'deploy activate' on different controllers
PASS: Remove a patch, running 'deploy start' and
      'deploy activate' on different controllers

Depends-on: I2a68091704518a4cae8507351c23e4a84d1ea027

Story: 2010676
Task: 51298

Change-Id: I282bdf8acb773aa724c710acc96f487bdfb8968d
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>