Restrict ceph mon hook execution
This commit adds a condition to run the ceph mon removal hook
only when rolling back to stx-8, which is the only scenario
where it is needed, and it causes issues if ran in other upgrade
paths. This commit also adds a TODO reminder to remove this hook
in the future when it is not needed anymore.

Test Plan (only affects AIO-DX)
PASS: AIO-DX - stx-10 -> stx-8 rollback successful
PASS: AIO-DX - stx-11 -> stx-10 rollback successful

Story: 2010676
Task: 51307

Change-Id: I5447deaa202fd78e983be26a9a230848bc2bc04e
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>