Enhance way to check if host is updated
Today PatchAgent checks if the host is updated by comparing the last
feed commit ID and the last sysroot commit ID, but this is not
accurate, since the sysroot commit can be not applied in the node.

This commit enhances this check by comparing the last feed commit ID
with the last deployment commit ID (it can be pending or active).
With this change, the system will be able to identify if the
deployment was correctly applied. If not, it will try to reapply.

Test-plan:
PASS: 'deploy start' a patch and see if the nodes report the mismatch
      with the deployment
PASS: 'deploy start' a major release and see if the nodes report the
      mismatch with the deployment
PASS: Apply/remove a patch in a DC lab with success
PASS: Apply a full major upgrade in a DX system with sucess

Story: 2010676
Task: 51322

Change-Id: I4658092da8a34d63cea6413981ce9f06aa0c0c16
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>