Invert order of undeploying inactive deployments
There is an issue on the logic of the undeploy_inactive_deployments
function when there are more than 2 deployments in the sysroot ostree.
For example, when there are 3 deployments:

0 -> ostree-commit-a
1 -> ostree-commit-b
2 -> ostree-commit-c

When deployment 1 is deleted, deployment 2 becomes the 1, and when
the function tries to remove 2 it fails, since there is no such
deployment, similar to an index out-of-bounds failure.

This commit fixes the issue by changing the logic to undeploy
in the inverse order, i.e. from the oldest to the newest. In the
example above it will undeploy in the order 2 -> 1. This commit
also fixes an issues with wrong usage of lstrip(), that would
remove characters incorrectly from the commit-id if it started
with characters of the string 'debian'.

Test Plan
PASS: stx-8 -> stx-10 major release deploy delete successfully
      and verify the inactive deployments were removed
PASS: force a scenario where more than 2 deployments exists and
      validate the function behaves as expected without errors

Story: 2010676
Task: 51321

Change-Id: I6e78ce199e6904096cec4c1f9d212d1448a9852b
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>