Fail if update scripts do not sync
Since VIM starts the patch deployment in one controller and finishes
in the other, the patch update scripts must be synced. So if it
fails, it must return false, and then the sync_from_nbr function will
try to sync again.
This way it ensures the update scripts sync between controllers.

Instead of deleting the entire /etc/update.d folder, it deletes only
the scripts inside.

Test-Plan:
PASS: Install a patch in a DX system using VIM with success
PASS: Do not raise an exception if the update scripts
      folder does not exist in the first controller

Story: 2010676
Task: 51377

Change-Id: I4ae36499e2a092199c82d5742172f3777c0ddc74
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>