Fix the wrong apt-ostree package feed
When installing the Debian packages, sometimes the package feed path
was wrongly passed, using the default component "updates", instead of
the release version.
This commit fixes it by building and using the correct path for each
patch.

Test-plan:
PASS: Deploy a patch in a SX system
PASS: Deploy a patch in a DX system
PASS: Deploy a patch in an upgraded SX system
PASS: Deploy a patch in an upgraded DX system

Story: 2010676
Task: 51379

Change-Id: I07aabdf49dae8474678a020917aaca5b57661c1d
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>