Fix patching output message
This commit sets the same output message for a major release upgrade
and a patching operation.

Test Plan:

PASS: Run a software deploy start on a system uploaded with the
latest patch and observe the message. Run the software deploy start
on a N-1 system uploaded with the current version iso. Verify that
the messages are the same.

Closes-bug: 2089412

Change-Id: I52e81bd588e0b535d358a5e6c624f5feb80341d9
Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>