Retry ostree deployment
Sometimes the "ostree admin deploy" command does not work in the
first try, we don't know yet the root cause of it. When this happens
the node trying to deploy, fails, reboots and tries to deploy again.
But this reboots cost a lot of time.

This commit identifies the fail in the "ostree admin deploy" command
earlier and retry without the need to reboot.

Test-Plan:
PASS: Install a patch in a SX system
PASS: Install a patch in a DX system
PASS: Install a major release in a DX system
PASS: Install a patch in a DC system
PASS: Install a patch in a system with nodes
      failing to install on the first try

Story: 2010676
Task: 51447

Change-Id: I8956dfd058394fc9a5927d08185fe8d26735aaa1
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>