Fix sysroot not match feed for major rollback
The software agent uses the latest_feed_commit value to check if the
correct deployment was installed in the node. The query function is
responsible for updating this value. But for major releases,
sometimes it is not possible to get this value updated, so query has
a major_release parameter to return early if it is a major release.

This commit fixes a query call that was not passing the major release
parameter. And resets the latest_feed_commit value in case it is a
major release.

Test-Plan:
PASS: Install a patch
PASS: Rollback a patch
PASS: Install a major release
PASS: Rollback a major release

Story: 2010676
Task: 51460

Change-Id: I8c9eeab39484bab1a4e8886deacc8fe32c1cf611
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>