Fix wrong ostree config file
Ostree config file must have a min-free-space-percent input into the
core (first) section. Assuming only core section existed at this
point, it was always adding this line at the end of the config file,
thus the end of core. But that is not the case anymore.

This commit, uses configParser to always put min-free-space-percent
into the core section.

Test-Plan:
PASS: Upload a pre-patched 24.09.1 ISO in a 22.12 system  and check
      if min-free-space-percent was put under core in the 24.09
      ostree config file.

Story: 2010676
Task: 51486

Change-Id: Ib180a748b17fbe1079a8b8bd424cc3312b6d0fcb
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>