Start software controller service after upgrade
When upgrading from 22.12 to a pre-patched ISO the
software-controller service is not being initialized.
This service among other things is responsible for initializing
apt-ostree and syncing some folders. Without it, the system is unable
to apply a patch after the upgrade.

This commit fixes it by starting the service after a major upgrade

Test-plan:
PASS: Upgrade a system from 22.12 to 24.09.1 and check if the service
      was started, apt-ostree initialized and folders synced.
PASS: Apply a patch with success

Story: 2010676
Task: 51516

Change-Id: I8ab8664ab309fc9dc5edf0973c220874b0692bc7
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>