Add nodetype check for service preset
This commit is to add the nodetype check during USM initialize
service to ensure the controller preset only exists in AIO controller.

This commit is to fix the issue where in AIO + Worker system, the
controller preset is incorrectly existing in worker node.

Test Plan:

PASS: build and deploy iso
PASS: ensure the controller preset not present in worker node
          in AIO+Worker system

Task: 51528
Story: 2010676

Change-Id: I827fa091ccf10f8e8d1c43f1122cd08ab1b40e78
Signed-off-by: junfeng-li <junfeng.li@windriver.com>