Clear leftover 900.001 alarm
The Patch in Progress alarm is a leftover of legacy patching and
is normally cleared during upgrade to usm releases by querying
the software API and the sw-patch API at the same time.

Corner cases were observed where patch-alarm-manager could not
communicate with one of the APIs and the alarm was not cleared,
carrying it over to the new release.

This change checks if the 900.001 alarm is raised and clears it
as it's no longer relevant on usm releases.

Test-Plan:
PASS: - Edit patch_alarm_manager.py and restart service
      - Check if alarm is cleared

closes-bug: 2090870

Change-Id: I633c590b2457e7fbe7abbefe934b07fbeb5356aa
Signed-off-by: mmachado <mmachado@windriver.com>