Create tombstone to fix ostree error in removal
When using a pre-patched ISO, the ostree history in the feed only
shows the last commi, the oldest commits are replaced by a message:
"History beyond this commit not fetched".

In these cases, ostree has a bug if we add a commit on this repo and
remove it later, the ostree complains a tombstone file is missing.
To fix this, after a patch removal, this review checks if this file
exists. If not the file is created.

It also creates a tombstone-commits=true in the repo config file, in
case in the future this bug is fixed on ostree side, then ostree will
be aware a tombstone file exists.

Test-Plan:
PASS: Remove a patch on a pre-patched system with success
PASS: Apply 2 patches and remove the last one on a pre-patched
      system with success
PASS: Check if tombstone-true was added in the config file

Closes-bug: 2098891

Change-Id: Icff3f834b81dd399736f14965e4b90b2130cce2e
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>