Fix minor and prestaging conflict
In a prepatched ISO, the metadata only shows the last patch and
ommits the rest. e.g.: A prepatched ISO 24.09.1, shows only 24.09.1,
but ommits 23.09.0.
In this case we can apply patch 24.09.2 (that requires 24.09.1)
without problem. But we cannot apply patch 24.09.100, because it
requires 24.09.0 (this is applied together with the prepatched ISO,
however the system does not know it).

This commit fixes it by reading a new field from the prepatched ISO
metadata, the <preinstalled_patches>. It is a list that contains all
the patches ID from a given prepatched ISO. e.g.: For 24.09.2
prepatched ISO, preinstalled_patches will be:
["component-24.09.0", "component-24.09.1"]

Using this list now the system is able to
apply patches from a different branch.

This commit also fixes the same problem during precheck.

Test-Plan:
PASS: Apply a prestaging patch (e.g. 24.09.100)
      in a minor prepatched ISO (e.g. 24.09.1)
PASS: Apply multiple patches
      (e.g. apply 24.09.200 after above test case)

Closes-bug: 2104065

Change-Id: Ia4b0ec7ec62d9de86f78a5dc398f1cf3bf95a1f9
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>