Recreate controller-feed remote after deploy host
For non-controller hosts, there is an additional ostree remote
called controller-feed that points to the ostree feed in the
active controller host. This remote needs to be updated locally
after a host is deployed with the to-release in upgrade scenario,
otherwise it will keep pointing to the from-release feed and cause
issues during the upgrade, leading the non-controller hosts to
mistakenly rollback to the from-release.

This commit uses an agent hook to delete the controller-feed
remote on these affected hosts, so that after the deploy host
and reboot the remote is recreated pointing to the right feed
on the active controller.

Test Plan
PASS: STANDARD - deploy non-controller hosts, verify the
      remote is updated accordingly and that the host is running
      to-release ostree commit-id after unlock
PASS: STANDARD - verify the same for previous test case but
      on the rollback scenario

Story: 2011357
Task: 51725

Change-Id: I1141e2a905f24eae33389130c5dd7929e925ad63
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>