SM service logs are being duplicated into postgres.log
The logging configuration update for the SM service logs
is invalid and conflicts with the postgres logging
facility (local0), causing the SM logs to appear in
both sm-service.log and postgres.log. Reverting to
facility (local3) for SM service logs to fix the
issue and adding upgrade and rollback support for
the syslog changes introduced.
Change that introduced the bug - https://review.opendev.org/c/starlingx/ha/+/923772

Test Plan: Verified the changes on simplex and duplex configurations.

PASS: AIO-Simplex - Verified service logs are written to
                    sm-service.log.Verfied that service
                    logs are not seen in sm.log
                    and postgres.log
PASS: AIO-Duplex -  Verified service logs are written to
                    sm-service.log.
                    Verified that service logs are not in sm.log and
                    postgres.log
PASS: AIO-Duplex -  Verified that service logs are written to
                    sm-service.log file after a controlled swact and
                    not seen in sm.log and postgres.log
PASS: AIO-Duplex -  Verified that upgrade from 22.12 to 24.09 is
                    successful and the logs are written to
                    sm-services.log file
PASS: AIO-Duplex -  Verified that rollback from 24.09 to 22.12 is
                    successful and logs are written to sm.log file

Closes-Bug: 2083632
Change-Id: If98be112c7509ebbd9dcd86c1e34c577d2108695
Signed-off-by: Sandhya Kalisetty <sandhya.kalisetty@windriver.com>