Fix pod accessibility issue after deploy host upgrade process
During the deploy host upgrade process, once controller-0
and controller-1 are deployed, the pods scheduled on
controller-1 become inaccessible. Kubernetes attempts to
create new pods but fails, while the old ones remain in
a non-functional state. This results in the following error:

"error: unable to upgrade connection: pod does not exist"

The root cause is unclear, but it is suspected to be related
to a network issue on controller-1. Restarting the
kube-apiserver resolves the issue, allowing the pods to
become accessible again.

Test Plan
PASS: AIO-DX - All pods are in a running state and kubectl exec
      is successful after deploy host is done.

Closes-Bug: 2103822

Change-Id: I9c9e3a7fbd62a7eb55111d2130b14012c2515736
Signed-off-by: Saba Touheed Mujawar <sabatouheed.mujawar@windriver.com>