Handle deployment interruption
This commit is an enhancement for handling interruption
during the USM upgrade.

When the USM upgrade is in the stage of deploy-start,
deploy-host, activate or activate-rollback,  if the interruption
occurs, such as host reboot, the deploy state will be set to
failed for the recent stage. After setting to failure state, the
USM upgrade can be re-tried.

Test Plan:

PASS: build and deploy iso
PASS: SX upgrade with deploy start interruption
PASS: DX upgrade with deploy start interruption

Task: 2011357
Story: 51849

Change-Id: I37341d9be5c17d1da7161e08c7b46fd86f28f589
Signed-off-by: junfeng-li <junfeng.li@windriver.com>