Use * for users with no password in /etc/shadow
CIS Benchmark considers the characters * and ! in the password field in
the /etc/shadow file to indicate that a user does not have a password
and is unable to log in if prompted for a password.

This replaces the character 'x' in some of those users with a '*' so the
benchmark can skip those users when checking some password-related
settings.

Test Plan:
PASS: Run build-pkgs -c -p software.
PASS: Run build-image.
PASS: Perform an upgrade in an AIO-SX and check the /etc/shadow file
      does not contain 'x' in the password field for any user.
PASS: Perform an upgrade in an AIO-DX and check the /etc/shadow file
      does not contain 'x' in the password field for any user.

Story: 2011283
Task: 51913

Change-Id: I0d31c98208ed904a3cedfe53ce45dfc842761c46
Signed-off-by: Rodrigo Tavares <Rodrigo.DosSantosTavares@windriver.com>