Configure code to use log format from software.conf
This commit stardandizes the log format to be fetched from the
software.conf file, it also have a check to use a default one in case
the log format is not present in the file which can happen in a 24.09
to 25.09 upgrade.

PASS: Performed and upgrade from stx10 -> stx11 and verified the logs.
PASS: Fresh install of stx11 and verified the logs.
PASS: Check software.conf have logging_default_format_string key
and value.

Story: 2011357
Task: 51915

Change-Id: Ied360551c9fc59c0e4f474ed94d1ba4a64065bd3
Signed-off-by: Luis Eduardo Bonatti <luizeduardo.bonatti@windriver.com>