Fix wrong order of error messages in the log
This commit is to fix the wrong order of error messages printed
in the log file. The wrong order was caused by the set() function
that is used to remove duplicates.

By removing duplicates, the python  printed table border is removed
as well because it is considered as duplicate.  Therefore, we'd keep
the duplicates and keep the printed table intact.

Test Plan:

PASS: build and deploy iso
PASS: run error activate script to verify output

Task: 51919
Story: 2011357
Change-Id: I918d2fe375b1fd782024737bf43981776c63e71a
Signed-off-by: junfeng-li <junfeng.li@windriver.com>