Add snapshot option to deploy precheck and start
This commit adds the --snapshot option to the deploy precheck
and start commands, to enable the lvm snapshot feature that will
be integrated with the USM upgrades in a future commit.

The deploy-precheck script is also changed, and a placeholder
message is added to it's output, to be replaced by the real
health checks in a separate commit.

Test Plan
PASS: verify deploy [precheck|start] help shows the new option
PASS: run deploy precheck with the new option (precheck support
      for the new option is part of another commit)
PASS: run deploy start with the new option, verify the deployment
      is created with the new option set as True
PASS: run deploy start without the new option, verify the deployment
      is created with the new option set as False

Story: 2011357
Task: 51928

Change-Id: Id7996894601de3331853d5122b07e453d2f2fbcc
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>