Remove unused prestaged releases
Prestage copies all releases deployed in the system controller to the
subclouds, but the user can apply an intermediate release in the
subcloud. Because of that, this review removes the unused commits and
releases, preventing the system from installing the wrong patch.

Test-Plan:
PASS: Prestage 24.09.100 and 24.09.200 and
      apply only 24.09.100 in the subcloud
PASS: Prestage 24.09.1 and 24.09.2 and
      apply only 24.09.1 in the subcloud
PASS: Prestage 24.09.100, 24.09.200 and 24.09.300
      and apply all of them in the subcloud
PASS: Prestage 24.09.1, 24.09.2 and 24.09.3
      and apply all of them in the subcloud
PASS: Remove applied patches from the subcloud
      24.09.100, 24.09.200 and 24.09.300
PASS: Remove applied patches from the subcloud
      24.09.1, 24.09.2 and 24.09.3

Closes-Bug: 2110536

Change-Id: I0699f65f1d679bafdb0bb98059aab86f14d40f84
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>