Replace etcd cloning for symlink
Previously, the etcd to-release directory was a clone of the
from-release database; with the current code a symlink is
created instead, so the etcd database is the same over all the
upgrade procedure, and the symlink is removed and from-release
directory is renamed to to-release during deploy delete.

This change is proposed to fix an issue occuring in multinode
systems where there is a mismatch between etcd (kubectl) and
the pods effectively running in a host (crictl), ultimately
leading to deploy activate failures.

Test Plan
PASS: AIO-SX e2e stx-10 -> stx-11 upgrade
PASS: AIO-DX e2e stx-10 -> stx-11 upgrade
PASS: AIO-DX orchestrated stx-10 -> stx-11 upgrade

Closes-bug: 2111588

Change-Id: I19bdffbbe7325e3edd9c45751dcac4af66acdf97
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>