Update kernel.env after creating ostree deployment
When an ostree deployment is created, it uses a kernel.env
from the ostree environment which does not contain local
changes such as selecting a real time kernel.

This change add a post-deployment step to update the kernel.env
using a utility method from puppet to ensure that the proper
kernel is selected, and that other values are preserved from
previous deployments.

Note:  Not all patches encounter this issue.
 This problem would only be seen if the patch introduced a
new directory under /boot/ostree

Test Plan:
 PASS:
 - Unlock a AIO-DX RT system.
 - lock / host-install a reboot-required patch / unlock.
 -- observe the system is running RT mode after reboot.
 -- observe that the /boot/ostree was different
 PASS:
 - Remove a kernel patch
 -- observe the system is still running RT mode after reboot

Closes-Bug: 2118552

Change-Id: Iea4d8e5e91e288715f531c032ff28dce915d9958
Signed-off-by: Soares Vieira <Lindley.Vieira@windriver.com>