Tombstone to fix ostree error in prepatched rehome
When using a prepatched ISO, the ostree history in the feed only
shows the last commit, the oldest commits are replaced by a message:
"History beyond this commit not fetched".

When this is pulled to a second repo (sysroot) and this repo needs
to delete some commit before pull, it will list all commits from the
original repo to check what must be deleted. During this list, if
it reaches the above message, it requires a tombstone commit to
avoid searching further and raise a file not found error.

Test-Plan:
PASS: Rehome a subcloud to a prepatched system controller, then
      apply a patch.
PASS: Remove the applied patch

Closes-Bug: 2113883

Change-Id: Iaa76ced6c2a66ae4b8666018150a2b74cea68177
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>