Add deploy delete check
This commit change is to ensure the from-release deploy data is
not deleted on the running system when upgrade failed.
The deploy delete operation is interrupted if the to-be-deleted
release is the same as running release.

Test Plan:

PASS: build and deploy the iso
PASS: delete the deploy before deploying the to-release on the system
PASS: delete the deploy after deploying the to-release on the system

Task: 52292
Story: 2011357
Change-Id: I6b74e381c8ab34a9b3a2d26f4e5c3529e5a9ce70
Signed-off-by: junfeng-li <junfeng.li@windriver.com>