Add start scripts to USM update framework
Adding patch scripts that run before and after
'software deploy start' command for each patch in the system controller.
This is an optional script, and can be added or not in a patch.

Test-Plan:
PASS: Run a patch without start scripts
PASS: Run a patch with start scripts
PASS: Check if scripts run in a subcloud
PASS: Apply multiple patches
PASS: Remove multiple patches

Story: 2010676
Task: 52438

Change-Id: Ie1c17f3b80ecacc9edaeea1ae9a5d6343d93e9d5
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>