Remove creation of a versioned etcd storage.
ETCD is no longer required to be upgraded as part of platform. This
commit changes the creation of a versioned etcd storage to a non
versioned one named 'db'. The new path is /opt/etcd/db.
This commit also changes the symlink between <release> and db to
a hardlink recursively.

Test Plan:
SX:
  PASS: Install 25.09.
  PASS: Upgrade 24.09 -> 25.09.
  PASS: Rollback after activate 25.09 -> 24.09.
  PASS: sw-deploy strategy 24.09 -> 25.09.
  PASS: Check crictl pods and kubectl get pods match results.
  PASS: Check Kubernetes Health Check Failed is not present.

DX:
  PASS: Install 25.09.
  PASS: Upgrade 24.09 -> 25.09.
  PASS: Rollback after activate 25.09 -> 24.09.
  PASS: sw-deploy strategy 24.09 -> 25.09.
  PASS: Check crictl pods and kubectl get pods match results.
  PASS: Check Kubernetes Health Check Failed is not present.

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/952490

Story: 2011357
Task: 52368

Change-Id: I99a07ec159a4794993e05b7c505de82a5b5e2372
Signed-off-by: Luis Eduardo Bonatti <luizeduardo.bonatti@windriver.com>