Upgrade hook for kernel param rcutree.kthread_prio
During upgrade the rcutree.kthread_prio kernel
parameter may be missing.
For lowlatency kernel upgrades if the kernel
parameter is not set we set it to default value
rcutree.kthread_prio=21

if it already set leave it as it is.
if the upgrade is standard kernel don't set it

Test plan:

PASS - AIO-SX: iso install
PASS - AIO-SX: check for double reboot on upgrade
               - standard kernel upgrade
               - lowlatency kernel upgrade

Story: 2011401
Task: 52394

Change-Id: I1808e0b066f770487c05b6d8b5e63ace8060e0f3
Signed-off-by: Kyale, Eliud <Eliud.Kyale@windriver.com>