Add service parameter apply activation script
Add a upgrade script to perform service parameter apply during
upgrade activation. This will be used to apply the change of the
securePort of kube-apiserver from 6443 to 16443.

It's also required to apply the change made by the script
"02-k8s-disable-sched-controllermanager-leader-election.sh". This
script is making a change in service parameters but doesn't apply it.

Test plan:
PASS: Deploy AIO-SX.
PASS: Upgrade AIO-SX from stx 10 to stx 11.

Story: 2011399
Task: 52178

Depends-on: https://review.opendev.org/c/starlingx/stx-puppet/+/954069

Change-Id: Ic8261ba089199809c77fc2b0c8ffb649a0aad91f
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>