VIM database restore from snapshot
During rollback using snapshot, after host unlock, VIM
DB is restored with the value it had at the time of taking
snapshot (before deploy start). This leads to issues as the
rollback strategy is not present in DB after unlock.

This change replaces VIM DB in snapshot with the active DB,
by dumping and loading the content to mounted snapshot. The
configuration file is updated so that after activate rollback,
the DB in snapshot acts as the current DB and all the changes
are written to it. After unlock when the snapshot DB is restored
it has the current values, therefore rollback strategy continues
as expected.

TEST PLAN:
Pass: Verify major upgrade rollback with snapshot on AIO-SX
      with VIM
Pass: Verify major upgrade rollback without snapshot on AIO-SX
      with VIM and USM
Pass: Verify snapshot is not mounted when doing rollback with
      USM

Story: 2011357
Task: 52483
Change-Id: I7d0789e0272aa43e18dd5de55656f046f0d381d9
Signed-off-by: sshathee <shunmugam.shatheesh@windriver.com>