reintroduce etc-merger
The new EtcMerger agent hook is to provide capability to delivery N+1
release config file that overwrites existing N release config file.
The hook copies the etc config files that are installed in below path
on N+1 load,
/usr/share/starlingx/usm/etc/
to the post ostree-merged etc directory in the path relative to /etc,
regardless the /etc config file was previously modified or not.

In a separated task, /etc directory will be backup prior to deploy N+1
to the host. Also selective etc config files will be restored from the
backup during a deploy host rollback.

Story: 2011357
Task: 52504

TCs:
    passed: SX upgrade, observed files from /usr/share/starlingx/usm/etc
            are copied over to /etc directory overwriting the modified
            /etc config file in the with relative path.

Change-Id: I76414222055ecc51cbb133d10d9f19b5214b9bfc
Signed-off-by: Bin Qian <bin.qian@windriver.com>