Fix in the optional extra.tar folder
Extra.tar is an optional folder in a patch. The way it checks
if this folder is present is rasing an uncaught exception.
This commit changes the way to check if it exists or not.

Test-Plan:
PASS: Upload a patch with extra.tar should succeed
PASS: Upload a patch without extra.tar should succeed

Closes-Bug: 2116956

Change-Id: I7708aa3d8a5577d2333ed2b8068275b8d7c661a0
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>