Fix the patch activation scripts folder
The /etc/update.d folder contains activation scripts that are synced
between the controllers at swact time. This folder was only being
created during the deploy start step, but if for some reason a swact
occurs before that, the sync was failing because the folder did not
exist.

This commit fixes it by creating this folder at the PatchController
initialization, allowing a swact to occur at any time without error.

Test-Plan:
PASS: Apply a patch with activation scripts and
      see if it's being synced with the other controller
PASS: Apply a patch without activation scripts,
      and the sync must not fail
PASS: Swact controllers prior to deploy start,
      and sync must not fail

Closes-Bug: 2117272

Change-Id: I7a3d178a8808a6922ae86bccdd12f50a56ba0613
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>