Fix start scripts parse
The start scripts receive an operation argument indicating if the patch
is being applied or removed. (e.g. --operation=remove)
But in some scripts the parse was not considering the "=" sign, causing
the script to wrongly run the defaut applying part even when removing.
This commit fixes the parse, aligning with the expect argument format

Test-Plan:
PASS: Apply a patch and check in the logs
      if restart sript runned in apply mode
PASS: Remove a patch and check in the logs
      if restart script runned in removal mode

Closes-Bug: 2117403

Change-Id: Ib72ab18cf7eb64a0b141f9e110fb8ef23cd23270
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>