Fix sync feed dir between controllers
During a deploy, the controllers were trying to sync all feed
directories, previous N-1, actual N, and future N+1 releases.
It must not sync the previous releases to avoid further data
transfer and errors.

This commit changes it by skipping the previous releases when
looping through all feed dir releases.

Test-Plan (in all cases, check if the feed repo was synced after swact):
PASS: Apply a patch in a DX system pre-bootstrap with success
PASS: Upgrade from 24.09 to 25.09 in a DX system with success
PASS: With an inactive build loaded, apply a patch in a DX system

Closes-Bug: 2118576

Change-Id: I3367aacffe6798ebf21c9a5abc2d7c481d2a7bf2
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>